+++ packages/core/src/enums/DepthTextureMode.ts
